Remove Coingi Exchange link as it is a fake exchange and stealing BTC.
Hello I have followed NMC for a while and really appreciate all that you are doing for the good of humanity. I would like to inform you that I followed the link to coingi Exchange and was promptly bilked out of the BTC I deposited and traded for NMC never saw a thing.